+++ src/main/java/com/buaa/compilec0/symbol/VariableSymbol.java
@@ -5,7 +5,7 @@ import com.buaa.compilec0.util.Pos;
 public class VariableSymbol extends Symbol{
     boolean initialized;
 
-    public VariableSymbol(SymbolType symbolType, DataType dataType, String symbolName, int level, int offset, Pos startPos, boolean initialized) {
+    public VariableSymbol(SymbolType symbolType, DataType dataType, String symbolName, int level, long offset, Pos startPos, boolean initialized) {
         super(symbolType, dataType, symbolName, level, offset, startPos);
         this.initialized = initialized;
     }
